backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.0
python-dotenv>=1.0

# MongoDB
pymongo>=4.6
mongoengine>=0.27

# AI/ML
scikit-learn>=1.4
numpy>=1.26
pandas>=2.2
openai>=1.0
transformers>=4.30

# Additional utilities
Pillow>=10.0
redis>=5.0
requests>=2.31
beautifulsoup4>=4.12
python-multipart>=0.0.6